migrating

+++ src/float/statistical/NormDiff/pc/nmppsNormDiff_Inf_32f.c
@@ -3,8 +3,8 @@
 void nmppsNormDiff_Inf_32f(const float* pSrc1, const float* pSrc2, int len, float* pNorm)
 {
 	int i;
-	int dif;
-	int max=0;
+	float dif;
+	float max=0;
 	for(i=0; i<len; i++){
 		dif=pSrc1[i]-pSrc2[i];
 		if (dif<0)
